Set up integration tests from examples
