+++ .github/workflows/nav2-integration-tests.yaml
@@ -35,7 +35,7 @@ jobs:
         uses: actions/upload-artifact@v4
         with:
           name: free-fleet-examples
-          path: ${{ runner.temp }}/free-fleet-examples.tar
+          path: free-fleet-examples.tar
 
   integration-tests:
     needs: [build-free-fleet-docker-image]
@@ -65,11 +65,10 @@ jobs:
         uses: actions/download-artifact@v4
         with:
           name: free-fleet-examples
-          path: ${{ runner.temp }}
 
       - name: Load image
         run: |
-          docker load --input ${{ runner.temp }}/free-fleet-examples.tar
+          docker load --input free-fleet-examples.tar
           docker image ls -a
 
       - name: Start test fixture containers
